refactor(web): extract module-card derivation into a hook
Pull the identical module-card data derivation out of DataplaneModules and ModuleStrip into a shared useModuleCards hook.
