+++ web/src/hooks/index.ts
@@ -40,3 +40,6 @@ export { useRollingWindow } from './useRollingWindow';
 export { useListNavigation } from './useListNavigation';
 export { useLaneCardNavigation } from './useLaneCardNavigation';
 export { usePageContribution } from './usePageContribution';
+
+export { useModuleCards } from './useModuleCards';
+export type { ModuleCardData } from './useModuleCards';
