+++ web/src/pages/builtin/dashboard/DataplaneModules.tsx
@@ -1,14 +1,12 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import type { InstanceInfo } from '../../../api/inspect';
 import {
-    computeModulePipelineUsage,
     getModuleCardAgentUsage,
-    getModuleDescription,
     getModuleRoute,
-    normalizeModuleName,
     type AgentUsage,
 } from '../inspect/utils';
+import { useModuleCards } from '../../../hooks/useModuleCards';
 import { MemoryBar } from '../inspect/MemoryBar';
 import { fmtIEC } from '../inspect/formatters';
 
@@ -21,25 +19,7 @@ export interface DataplaneModulesProps {
 export const DataplaneModules: React.FC<DataplaneModulesProps> = ({ instance, usage }) => {
     const navigate = useNavigate();
     const modules = instance.dp_modules ?? [];
-    const configs = instance.cp_configs ?? [];
-
-    const pipeUsage = useMemo(() => computeModulePipelineUsage(instance), [instance]);
-
-    const moduleData = useMemo(
-        () =>
-            modules.map((m, idx) => {
-                const name = m.name ?? '';
-                const key = name || `module-${idx}`;
-                const moduleKey = normalizeModuleName(name);
-                const cfg = configs.filter(
-                    (c) => normalizeModuleName(c.type ?? '') === moduleKey,
-                ).length;
-                const pipe = pipeUsage.get(moduleKey) ?? 0;
-                const inUse = cfg > 0 || pipe > 0;
-                return { key, name, cfg, pipe, inUse, desc: getModuleDescription(name) };
-            }),
-        [modules, configs, pipeUsage],
-    );
+    const moduleData = useModuleCards(instance);
 
     const colCount = Math.min(9, Math.max(1, modules.length));
